feat: move editor in lazy loaded component
